feat: configure keycloak container
This is a general "basic" configuration, making the container ready for
use in a production environment and allowing access through Caddy.
Actual configuration of the Dansdata realm etc. still needs to be done.

The work in this commit is primarily based on the official
["Configuring Keycloak for production" guide](
  https://www.keycloak.org/server/configuration-production
).

I found [this](https://www.keycloak.org/server/all-config) page really
useful when looking up config options.